requirements.txt
streamlit
pandas
scikit-learn
openpyxl
google-cloud-storage
google-cloud-bigquery
db-dtypes